confirmation for address deletion added

+++ src/app/address-list/address-details/address-details.component.ts
@@ -35,7 +35,7 @@ export class AddressDetailsComponent implements OnInit {
     }
 
     onDeleteAddress(): void {
-        if (this.address) {
+        if (this.address && this.addressService.confirmDeletion(this.address.name)) {
             this.addressService.deleteAddress(this.address.id);
             this.router.navigate(['/']);
         }
